+++ R/ref-grid.R
@@ -525,6 +525,7 @@ ref_grid <- function(object, at, cov.reduce = mean, cov.keep = get_emm_option("c
     
     # we've added args `misc` and `options` so emm_basis methods can access and use these if they want
     basis = emm_basis(object, trms, xlev, grid, misc = attr(data, "misc"), options = options, ...)
+    environment(basis$dffun) = baseenv()   # releases unnecessary storage
     if(length(basis$bhat) != ncol(basis$X))
         stop("Something went wrong:\n",
              " Non-conformable elements in reference grid.",